+++ core/integration/setup_test.go
@@ -224,7 +224,6 @@ func (e *executionTestSetup) registerTimeServiceCallbacks() {
 		e.ntry.OnTick,
 		e.banking.OnTick,
 		e.delegationEngine.OnTick,
-		e.rewardsEngine.OnTick,
 		e.builtinOracle.OnTick,
 		e.stateVarEngine.OnTick,
 		e.executionEngine.OnTick,
@@ -250,20 +249,12 @@ func (e *executionTestSetup) registerNetParamsCallbacks() error {
 			Watcher: e.executionEngine.OnMarketFeeFactorsInfrastructureFeeUpdate,
 		},
 		netparams.WatchParam{
-			Param:   netparams.MarketLiquidityStakeToCCYSiskas,
+			Param:   netparams.MarketLiquidityStakeToCCYVolume,
 			Watcher: e.executionEngine.OnSuppliedStakeToObligationFactorUpdate,
 		},
 		netparams.WatchParam{
 			Param:   netparams.MarketValueWindowLength,
 			Watcher: e.executionEngine.OnMarketValueWindowLengthUpdate,
-		},
-		netparams.WatchParam{
-			Param:   netparams.MarketTargetStakeScalingFactor,
-			Watcher: e.executionEngine.OnMarketTargetStakeScalingFactorUpdate,
-		},
-		netparams.WatchParam{
-			Param:   netparams.MarketTargetStakeTimeWindow,
-			Watcher: e.executionEngine.OnMarketTargetStakeTimeWindowUpdate,
 		},
 		netparams.WatchParam{
 			Param:   netparams.MarketLiquidityProvidersFeeDistribitionTimeStep,
@@ -280,10 +271,6 @@ func (e *executionTestSetup) registerNetParamsCallbacks() error {
 		netparams.WatchParam{
 			Param:   netparams.MarketLiquidityBondPenaltyParameter,
 			Watcher: e.executionEngine.OnMarketLiquidityBondPenaltyUpdate,
-		},
-		netparams.WatchParam{
-			Param:   netparams.MarketLiquidityTargetStakeTriggeringRatio,
-			Watcher: e.executionEngine.OnMarketLiquidityTargetStakeTriggeringRatio,
 		},
 		netparams.WatchParam{
 			Param:   netparams.MarketAuctionMinimumDuration,