+++ wallet/api/admin_sign_transaction_test.go
@@ -394,7 +394,7 @@ func testAdminSigningTransactionWithMalformedTransactionFails(t *testing.T) {
 	})
 
 	// then
-	assertInvalidParams(t, errorDetails, api.ErrTransactionIsMalformed)
+	assertInvalidParams(t, errorDetails, api.ErrTransactionIsNotValidVegaCommand)
 	assert.Empty(t, result)
 }