+++ wallet/api/helpers_for_test.go
@@ -175,8 +175,8 @@ func dummyServiceShutdownSwitch() *api.ServiceShutdownSwitch {
 }
 
 var (
-	testTransactionJSON          = `{"voteSubmission":{"proposalId":"eb2d3902fdda9c3eb6e369f2235689b871c7322cf3ab284dde3e9dfc13863a17","value":"VALUE_YES"}}`
-	testMalformedTransactionJSON = `{"voteSubmission":{"proposalId":"not real id","value":"VALUE_YES"}}`
+	fakeTransaction          = `{"voteSubmission":{"proposalId":"eb2d3902fdda9c3eb6e369f2235689b871c7322cf3ab284dde3e9dfc13863a17","value":"VALUE_YES"}}`
+	fakeMalformedTransaction = `{"voteSubmission":{"proposalId":"not real id","value":"VALUE_YES"}}`
 )
 
 func transactionFromJSON(t *testing.T, JSON string) map[string]any {
@@ -188,12 +188,10 @@ func transactionFromJSON(t *testing.T, JSON string) map[string]any {
 
 func testTransaction(t *testing.T) map[string]any {
 	t.Helper()
-	return transactionFromJSON(t, testTransactionJSON)
+	return transactionFromJSON(t, fakeTransaction)
 }
 
 func testMalformedTransaction(t *testing.T) map[string]any {
 	t.Helper()
-	return transactionFromJSON(t, testMalformedTransactionJSON)
+	return transactionFromJSON(t, fakeMalformedTransaction)
 }
-
-var testEncodedTransaction = "ewogICAgInZvdGVTdWJtaXNzaW9uIjogewogICAgICAgICJwcm9wb3NhbElkIjogImViMmQzOTAyZmRkYTljM2ViNmUzNjlmMjIzNTY4OWI4NzFjNzMyMmNmM2FiMjg0ZGRlM2U5ZGZjMTM4NjNhMTciLAogICAgICAgICJ2YWx1ZSI6ICJWQUxVRV9ZRVMiCiAgICB9Cn0K"